requirements.txt
rpi.gpio
aliot-py